+++ packages/admin-host/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @voyant-travel/admin-host
 
+## 0.134.0
+
+### Patch Changes
+
+- @voyant-travel/admin-app@0.176.0
+
+## 0.133.0
+
+### Patch Changes
+
+- @voyant-travel/admin-app@0.175.0
+
 ## 0.132.0
 
 ### Patch Changes