+++ packages/cruises/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @voyant-travel/cruises
 
+## 0.239.2
+
+### Patch Changes
+
+- Updated dependencies [d3b17e2]
+  - @voyant-travel/finance@0.250.0
+  - @voyant-travel/catalog@0.256.1
+
+## 0.239.1
+
+### Patch Changes
+
+- Updated dependencies [a41a73a]
+  - @voyant-travel/catalog@0.256.0
+  - @voyant-travel/cruises-contracts@0.105.33
+
 ## 0.239.0
 
 ### Minor Changes