+++ packages/legal/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @voyant-travel/legal
 
+## 0.251.4
+
+### Patch Changes
+
+- Updated dependencies [d3b17e2]
+  - @voyant-travel/finance@0.250.0
+  - @voyant-travel/catalog@0.256.1
+  - @voyant-travel/commerce@0.51.2
+  - @voyant-travel/distribution@0.228.2
+  - @voyant-travel/relationships@0.134.3
+
+## 0.251.3
+
+### Patch Changes
+
+- Updated dependencies [a41a73a]
+  - @voyant-travel/catalog@0.256.0
+  - @voyant-travel/commerce@0.51.1
+  - @voyant-travel/distribution@0.228.1
+
 ## 0.251.2
 
 ### Patch Changes