+++ packages/mice-react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @voyant-travel/mice-react
 
+## 0.151.0
+
+### Patch Changes
+
+- @voyant-travel/relationships-react@0.283.0
+
+## 0.150.0
+
+### Patch Changes
+
+- @voyant-travel/relationships-react@0.282.0
+
 ## 0.149.0
 
 ### Patch Changes