+++ packages/operator-settings-react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @voyant-travel/operator-settings-react
 
+## 0.140.0
+
+### Patch Changes
+
+- Updated dependencies [d3b17e2]
+  - @voyant-travel/finance@0.250.0
+  - @voyant-travel/finance-react@0.283.0
+  - @voyant-travel/admin-app@0.176.0
+
+## 0.139.0
+
+### Patch Changes
+
+- @voyant-travel/finance-react@0.282.0
+- @voyant-travel/admin-app@0.175.0
+
 ## 0.138.0
 
 ### Patch Changes