+++ packages/products-contracts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @voyant-travel/products-contracts
 
+## 0.111.5
+
+### Patch Changes
+
+- Updated dependencies [a41a73a]
+  - @voyant-travel/catalog-contracts@0.134.0
+
 ## 0.111.4
 
 ### Patch Changes