+++ packages/storefront/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @voyant-travel/storefront
 
+## 0.257.2
+
+### Patch Changes
+
+- Updated dependencies [d3b17e2]
+  - @voyant-travel/finance@0.250.0
+  - @voyant-travel/catalog@0.256.1
+  - @voyant-travel/commerce@0.51.2
+  - @voyant-travel/cruises@0.239.2
+  - @voyant-travel/flights@0.237.8
+
+## 0.257.1
+
+### Patch Changes
+
+- Updated dependencies [a41a73a]
+  - @voyant-travel/catalog-contracts@0.134.0
+  - @voyant-travel/catalog@0.256.0
+  - @voyant-travel/flights@0.237.7
+  - @voyant-travel/commerce@0.51.1
+  - @voyant-travel/cruises@0.239.1
+
 ## 0.257.0
 
 ### Minor Changes