+++ packages/trips-react/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @voyant-travel/trips-react
 
+## 0.276.0
+
+### Patch Changes
+
+- Updated dependencies [d3b17e2]
+  - @voyant-travel/finance@0.250.0
+  - @voyant-travel/bookings-react@0.283.0
+  - @voyant-travel/flights-react@0.283.0
+  - @voyant-travel/catalog-react@0.281.0
+  - @voyant-travel/relationships-react@0.283.0
+
+## 0.275.0
+
+### Patch Changes
+
+- Updated dependencies [a41a73a]
+  - @voyant-travel/catalog-contracts@0.134.0
+  - @voyant-travel/catalog-react@0.280.0
+  - @voyant-travel/catalog@0.256.0
+  - @voyant-travel/bookings-react@0.282.0
+  - @voyant-travel/flights-react@0.282.0
+  - @voyant-travel/relationships-react@0.282.0
+
 ## 0.274.0
 
 ### Patch Changes